+++ tests/Integration/NestedGroupDuplicateKey/di-web.php
