+++ tests/Integration/NestedGroupDuplicateKey/di.php
