+++ tests/Integration/NestedGroupDuplicateKey/packages/a/composer.json
