+++ tests/Integration/NestedGroupDuplicateKey/packages/a/di.php
